+++ packages/sv/src/index.ts
@@ -3,9 +3,9 @@ import { create as _create, type Options as CreateOptions } from './create/index
 
 export type { TemplateType, LanguageType } from './create/index.ts';
 
+export function create(options: CreateOptions): void;
 /** @deprecated use `create({ cwd, ...options })` instead. */
 export function create(cwd: string, options: Omit<CreateOptions, 'cwd'>): void;
-export function create(options: CreateOptions): void;
 export function create(
 	cwdOrOptions: string | CreateOptions,
 	legacyOptions?: Omit<CreateOptions, 'cwd'>